% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_PHM.R
\name{tidy_PHM}
\alias{tidy_PHM}
\title{Tidy Public Health Metrics}
\usage{
tidy_PHM(x, primary_segment, secondary_segment)
}
\arguments{
\item{x}{A dataframe containing the raw data.}

\item{primary_segment}{A string representing the primary segment column name.}

\item{secondary_segment}{A string representing the secondary segment column name.}
}
\value{
A tidy dataframe.
}
\description{
This function transforms a messy public health metrics dataframe into a tidy format.
}
